Fixed issue with model flush keys and Invalidator.find_flush_lists() model flush keys were being prefixed with ":mflush" instead of ":flush". This was problematic because Invalidator.find_flush_lists() does the reverse lookup of flush keys by running a map on object and query keys through caching.invalidation.flush_key(). A brief summary of the process:
 - caching.base.CachingQuerySet.iterator()
 - caching.base.CacheMachine.__iter__()
 - caching.base.CacheMachine.cache_objects(objects)
 - caching.invalidation.Invalidator.cache_objects(objects)
 - flush_keys += [o.model_flush_key() for o in objects]

where model_flush_key is
    return caching.invalidation.model_flush_key(self.model_key)

So, for a model blog.models.Post, our values are:
    model_key       = m:blog.post
    model_flush_key = :mflush:dee0a310068e1eacfa46d0c10c4bf3ca
        (the latter is the md5 digest of "m:blog.post")

but when "m:blog.post" is passed to Invalidator.find_flush_lists(),
it does a lookup for ":flush:dee0a310068e1eacfa46d0c10c4bf3ca",
which does not return any results, and so nothing is flushed.

This commit removes caching.invalidation.model_flush_key() and has
the relevant functions in caching.base use flush_key() instead,
which gives it the normal ":flush" prefix so that the reverse
lookup for the model key succeeds.