Use obj._meta.db_table for object cache key, not obj._meta.__unicode__()
obj._meta.__unicode__() will differ for the same object if it has
deferred fields. For instance, assume a model blog.models.Post. The
object returned from the query:

    posts = Post.objects.filter(pk=1)

will have the cache key "o:blog.post:1".

Because django requires deferred models to be registered in the app
cache separately, they are given a separate name, and so return a
different result when passed through unicode(). So the cache key for the
object returned from:

    posts = Post.objects.filter(pk=1).defer("content")

is "o:blog.post_deferred_content:1".

Consequently, the latter queryset will not get registered with the
object's flush key, and won't get cleared when the object changes.